+++ dominode/stac_endpoint/management/commands/index_item.py
@@ -2,11 +2,11 @@ from typing import Any
 from django.core.management.base import BaseCommand
 import json
 from ...minio_client.crawl_collection import CrawlAllCollections
+from ...minio_client.map_catalog import MapIndex
 import pystac
 import os
 from ...helpers import createStacItem
 from ...minio_client.client import client
-import pprint
 
 
 class Command(BaseCommand):
@@ -20,35 +20,49 @@ class Command(BaseCommand):
 
             collection_name = collection["collection_name"]
             itemsArr = collection["items"]
-            
-            collection_items = []
+            itemsNotIndexed = []
 
-            for folder in itemsArr:
+            for item in itemsArr:
+                item_name = item["item_name"]
+                map_index = MapIndex()
+                index = map_index.mapItem(collection_name, item_name)
 
-                collection_item = createStacItem(assets=folder["assets"], item_name=folder["item_name"], collection_name=collection_name)
-                collection_items.append(collection_item)
+                if len(index) > 0:
+                    self.stdout.write('index.json exists')
+                else:
+                    itemsNotIndexed.append(item_name)
                 
-            collection = pystac.ItemCollection(collection_items)
-            collection_dict = collection.to_dict()
-            features = collection_dict["features"]
-            pp = pprint.PrettyPrinter(width=41, compact=True)
-            
-            for feature in features:
-                item_name = feature["id"]
-                _collection_name = feature["collection"]
+            self.stdout.write(f"items not indexed {itemsNotIndexed}")
 
-                _full_path = os.path.realpath(__file__)
-                _dir_path = (os.path.dirname(_full_path))
-                _file_path = f'{_dir_path}/tmp/item/index.json'
-                self.stdout.write(str(_file_path))
+            if len(itemsNotIndexed) > 0:
+                collection_items = []
 
-                with open(_file_path, 'w') as fp:
-                    json.dump(feature, fp, indent=4)
+                for folder in itemsArr:
 
-                client.fput_object(
-                    "dominode", f"stac/{_collection_name}/items/{item_name}/index.json", _file_path
-                )
+                    if folder["item_name"] in itemsNotIndexed:
+                        collection_item = createStacItem(assets=folder["assets"], item_name=folder["item_name"], collection_name=collection_name)
+                        collection_items.append(collection_item)
+                    
+                collection = pystac.ItemCollection(collection_items)
+                collection_dict = collection.to_dict()
+                features = collection_dict["features"]
+                
+                for feature in features:
+                    item_name = feature["id"]
+                    _collection_name = feature["collection"]
 
-                os.remove(_file_path)
+                    _full_path = os.path.realpath(__file__)
+                    _dir_path = (os.path.dirname(_full_path))
+                    _file_path = f'{_dir_path}/tmp/item/index.json'
+                    self.stdout.write(str(_file_path))
+
+                    with open(_file_path, 'w') as fp:
+                        json.dump(feature, fp, indent=4)
+
+                    client.fput_object(
+                        "dominode", f"stac/{_collection_name}/items/{item_name}/index.json", _file_path
+                    )
+
+                    os.remove(_file_path)
 
         self.stdout.write('Item collection has been indexed successfully')
